pr-fix: complete merging with 'master'
