+++ src/Arcus.Messaging.Tests.Integration/Fixture/KeyVault.cs
@@ -1,4 +1,5 @@
-﻿using GuardNet;
+﻿using System;
+using GuardNet;
 
 namespace Arcus.Messaging.Tests.Integration.Fixture
 {
@@ -13,6 +14,8 @@ namespace Arcus.Messaging.Tests.Integration.Fixture
         /// <param name="vaultUri">The URI referencing the Azure Key Vault instance.</param>
         /// <param name="secretName">The name of the secret in the Azure Key Vault instance.</param>
         /// <param name="secretNewVersionCreated">The event endpoint of the Azure Key Vault 'Secret new version created' event.</param>
+        /// <exception cref="ArgumentException">Thrown when the <paramref name="vaultUri"/> or <paramref name="secretName"/> is blank.</exception>
+        /// <exception cref="ArgumentNullException">Thrown when the <paramref name="secretNewVersionCreated"/> is <c>null</c>.</exception>
         public KeyVault(string vaultUri, string secretName, KeyVaultEventEndpoint secretNewVersionCreated)
         {
             Guard.NotNullOrWhitespace(vaultUri, nameof(vaultUri));
@@ -38,24 +41,5 @@ namespace Arcus.Messaging.Tests.Integration.Fixture
         /// Gets the event endpoint of the Azure Key Vault 'Secret new version created' event.
         /// </summary>
         public KeyVaultEventEndpoint SecretNewVersionCreated { get; }
-    }
-
-    public class KeyVaultEventEndpoint
-    {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="KeyVaultEventEndpoint"/> class.
-        /// </summary>
-        /// <param name="connectionString">The connection string to connect to the endpoint that will handle the Azure Key Vault event.</param>
-        public KeyVaultEventEndpoint(string connectionString)
-        {
-            Guard.NotNullOrWhitespace(connectionString, nameof(connectionString));
-
-            ConnectionString = connectionString;
-        }
-
-        /// <summary>
-        /// Gets the connection string to connect to the endpoint that will handle the Azure Key Vault event.
-        /// </summary>
-        public string ConnectionString { get; }
     }
 }
